+++ examples/pool.rs
@@ -1,9 +1,11 @@
+//! Task pool example - this demonstrates running several async tasks concurrently.
+
 use async_std::task::sleep;
 use bevy::{app::PanicHandlerPlugin, log::LogPlugin, prelude::*};
-use bevy_async_task::{AsyncTaskPool, AsyncTaskStatus};
-use std::time::Duration;
+use bevy_async_task::AsyncTaskPool;
+use std::{task::Poll, time::Duration};
 
-fn system1(mut task_pool: AsyncTaskPool<u64>) {
+fn system1(mut task_pool: AsyncTaskPool<'_, u64>) {
     if task_pool.is_idle() {
         info!("Queueing 5 tasks...");
         for i in 1..=5 {
@@ -15,12 +17,13 @@ fn system1(mut task_pool: AsyncTaskPool<u64>) {
     }
 
     for status in task_pool.iter_poll() {
-        if let AsyncTaskStatus::Finished(t) = status {
+        if let Poll::Ready(t) = status {
             info!("Received {t}");
         }
     }
 }
 
+/// Entry point
 pub fn main() {
     App::new()
         .add_plugins((MinimalPlugins, LogPlugin::default(), PanicHandlerPlugin))