+++ src/lib.rs
@@ -1,22 +1,13 @@
-#![deny(missing_docs)]
 //! Ergonomic abstractions to async programming in Bevy for all platforms.
 
+mod error;
 mod receiver;
 mod task;
 mod task_pool;
 mod task_runner;
 
+pub use error::TaskError;
 pub use receiver::AsyncReceiver;
-pub use task::{AsyncTask, TimeoutError};
+pub use task::AsyncTask;
 pub use task_pool::AsyncTaskPool;
 pub use task_runner::AsyncTaskRunner;
-
-/// A poll status for an [`AsyncTask`].
-pub enum AsyncTaskStatus<T> {
-    /// No task is currently being polled.
-    Idle,
-    /// The task is currently working.
-    Pending,
-    /// The task is finished.
-    Finished(T),
-}