+++ src/task.rs
@@ -1,13 +1,21 @@
-use crate::AsyncReceiver;
+use crate::{AsyncReceiver, TaskError};
 #[cfg(not(target_arch = "wasm32"))]
 use async_compat::CompatExt;
 use async_std::future::timeout;
 use bevy::utils::{ConditionalSend, ConditionalSendFuture};
-use std::{future::pending, pin::Pin, time::Duration};
+use futures::task::AtomicWaker;
+use std::{
+    fmt::Debug,
+    future::pending,
+    pin::Pin,
+    sync::{
+        atomic::{AtomicBool, Ordering},
+        Arc,
+    },
+    task::Poll,
+    time::Duration,
+};
 use tokio::sync::oneshot;
-
-// Re-export timeout error
-pub use async_std::future::TimeoutError;
 
 /// A wrapper type around an async future. The future may be executed
 /// asynchronously by an [`AsyncTaskRunner`](crate::AsyncTaskRunner) or
@@ -18,30 +26,25 @@ pub struct AsyncTask<T> {
     receiver: AsyncReceiver<T>,
 }
 
+impl<T> Debug for AsyncTask<T>
+where
+    T: Debug,
+{
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("AsyncTask")
+            .field("fut", &"<future>")
+            .field("receiver", &self.receiver)
+            .finish()
+    }
+}
+
 impl<T> AsyncTask<T>
 where
     T: ConditionalSend + 'static,
 {
     /// Never resolves to a value or finishes.
-    pub fn pending() -> AsyncTask<T> {
-        AsyncTask::new(pending())
-    }
-
-    /// Add a timeout to the task.
-    pub fn with_timeout(mut self, dur: Duration) -> AsyncTask<Result<T, TimeoutError>> {
-        let (tx, rx) = oneshot::channel();
-        let new_fut = async move {
-            let result = timeout(dur, self.fut)
-                .await
-                .map(|_| self.receiver.try_recv().unwrap());
-            _ = tx.send(result);
-        };
-        let fut = Box::pin(new_fut);
-        let receiver = AsyncReceiver {
-            received: false,
-            buffer: rx,
-        };
-        AsyncTask::<Result<T, TimeoutError>> { fut, receiver }
+    pub fn pending() -> Self {
+        Self::new(pending())
     }
 }
 
@@ -53,54 +56,81 @@ impl<T> AsyncTask<T> {
         F::Output: ConditionalSend + 'static,
     {
         let (tx, rx) = oneshot::channel();
-        let new_fut = async move {
-            #[cfg(target_arch = "wasm32")]
-            let result = fut.await;
-            #[cfg(not(target_arch = "wasm32"))]
-            let result = fut.compat().await;
-            _ = tx.send(result);
+        let waker = Arc::new(AtomicWaker::new());
+        let received = Arc::new(AtomicBool::new(false));
+        let fut = {
+            let waker = waker.clone();
+            let received = received.clone();
+            async move {
+                #[cfg(target_arch = "wasm32")]
+                let result = fut.await;
+                #[cfg(not(target_arch = "wasm32"))]
+                let result = fut.compat().await;
+                if let Ok(()) = tx.send(result) {
+                    // Wait for the receiver to get the result before dropping.
+                    futures::future::poll_fn(|cx| {
+                        waker.register(cx.waker());
+                        if received.load(Ordering::Relaxed) {
+                            Poll::Ready(())
+                        } else {
+                            Poll::Pending::<()>
+                        }
+                    })
+                    .await;
+                }
+            }
         };
-        let fut = Box::pin(new_fut);
+        let fut = Box::pin(fut);
         let receiver = AsyncReceiver {
-            received: false,
-            buffer: rx,
+            received,
+            waker,
+            receiver: rx,
         };
         Self { fut, receiver }
     }
 
     /// Create an async task from a future with a timeout.
-    pub fn new_with_timeout<F>(dur: Duration, fut: F) -> AsyncTask<Result<T, TimeoutError>>
+    pub fn new_with_timeout<F>(dur: Duration, fut: F) -> AsyncTask<Result<T, TaskError>>
     where
         F: ConditionalSendFuture<Output = T> + 'static,
         F::Output: ConditionalSend + 'static,
     {
         let (tx, rx) = oneshot::channel();
-        let new_fut = async move {
-            #[cfg(target_arch = "wasm32")]
-            let result = timeout(dur, fut).await;
-            #[cfg(not(target_arch = "wasm32"))]
-            let result = timeout(dur, fut.compat()).await;
-            _ = tx.send(result);
+        let waker = Arc::new(AtomicWaker::new());
+        let received = Arc::new(AtomicBool::new(false));
+        let fut = {
+            let waker = waker.clone();
+            let received = received.clone();
+            async move {
+                #[cfg(target_arch = "wasm32")]
+                let result = timeout(dur, fut).await.map_err(TaskError::Timeout);
+                #[cfg(not(target_arch = "wasm32"))]
+                let result = timeout(dur, fut.compat()).await.map_err(TaskError::Timeout);
+                if let Ok(()) = tx.send(result) {
+                    // Wait for the receiver to get the result before dropping.
+                    futures::future::poll_fn(|cx| {
+                        waker.register(cx.waker());
+                        if received.load(Ordering::Relaxed) {
+                            Poll::Ready(())
+                        } else {
+                            Poll::Pending::<()>
+                        }
+                    })
+                    .await;
+                }
+            }
         };
-        let fut = Box::pin(new_fut);
+        let fut = Box::pin(fut);
         let receiver = AsyncReceiver {
-            received: false,
-            buffer: rx,
+            received,
+            waker,
+            receiver: rx,
         };
-        AsyncTask::<Result<T, TimeoutError>> { fut, receiver }
-    }
-
-    /// Block awaiting the task result. Can only be used outside of async
-    /// contexts.
-    pub fn blocking_recv(self) -> T {
-        let (fut, mut rx) = self.into_parts();
-        bevy::tasks::block_on(fut);
-        rx.buffer.try_recv().unwrap()
+        AsyncTask::<Result<T, TaskError>> { fut, receiver }
     }
 
     /// Break apart the task into a runnable future and the receiver. The
     /// receiver is used to catch the output when the runnable is polled.
-    #[allow(clippy::type_complexity)]
     #[must_use]
     pub fn into_parts(
         self,
@@ -118,7 +148,7 @@ where
     Fnc::Output: ConditionalSend + 'static,
 {
     fn from(value: Fnc) -> Self {
-        AsyncTask::new(value)
+        Self::new(value)
     }
 }
 
@@ -145,12 +175,6 @@ mod test {
             Ok(v) => assert_eq!(3, v),
             Err(e) => panic!("the sender dropped ({e})"),
         }
-    }
-
-    #[test]
-    fn test_blocking_recv() {
-        let task = AsyncTask::new(async move { 5 });
-        assert_eq!(5, task.blocking_recv());
     }
 
     #[tokio::test]
@@ -202,8 +226,12 @@ mod test {
             select! {
                 _ = (&mut fetch).fuse() => {
                     if let Some(v) = rx.try_recv() {
-                        assert!(v.is_err(), "timeout should have triggered!");
-                        break 'result;
+                        if matches!(v, Err(TaskError::Timeout(_))) {
+                            // Good ending
+                            break 'result;
+                        } else {
+                            panic!("timeout should have triggered!");
+                        }
                     } else {
                         // Reset the clock
                         fetch.reset(Duration::from_millis(1));
@@ -216,7 +244,7 @@ mod test {
 
     #[tokio::test]
     async fn test_with_timeout() {
-        let task = AsyncTask::<()>::pending().with_timeout(Duration::from_millis(5));
+        let task = AsyncTask::new_with_timeout(Duration::from_millis(5), pending::<()>());
         let (fut, mut rx) = task.into_parts();
 
         assert_eq!(None, rx.try_recv());
@@ -232,8 +260,12 @@ mod test {
             select! {
                 _ = (&mut fetch).fuse() => {
                     if let Some(v) = rx.try_recv() {
-                        assert!(v.is_err(), "timeout should have triggered!");
-                        break 'result;
+                        if matches!(v, Err(TaskError::Timeout(_))) {
+                            // Good ending
+                            break 'result;
+                        } else {
+                            panic!("timeout should have triggered!");
+                        }
                     } else {
                         // Reset the clock
                         fetch.reset(Duration::from_millis(1));
@@ -271,13 +303,9 @@ mod test {
             Ok(JsValue::NULL)
         }))
         .await
-        .unwrap();
-    }
-
-    #[wasm_bindgen_test]
-    fn test_blocking_recv() {
-        let task = AsyncTask::new(async move { 5 });
-        assert_eq!(5, task.blocking_recv());
+        .unwrap_or_else(|e| {
+            panic!("awaiting promise failed: {e:?}");
+        });
     }
 
     #[wasm_bindgen_test]
@@ -293,7 +321,9 @@ mod test {
             Ok(JsValue::NULL)
         }))
         .await
-        .unwrap();
+        .unwrap_or_else(|e| {
+            panic!("awaiting promise failed: {e:?}");
+        });
 
         // Spawn
         assert_eq!(Some(5), rx.try_recv());
@@ -312,16 +342,20 @@ mod test {
             Ok(JsValue::NULL)
         }))
         .await
-        .unwrap();
+        .unwrap_or_else(|e| {
+            panic!("awaiting promise failed: {e:?}");
+        });
 
         // Spawn
-        let v = rx.try_recv().expect("future loaded no value");
+        let v = rx.try_recv().unwrap_or_else(|| {
+            panic!("expected result after await");
+        });
         assert!(v.is_err(), "timeout should have triggered!");
     }
 
     #[wasm_bindgen_test]
     async fn test_with_timeout() {
-        let task = AsyncTask::<()>::pending().with_timeout(Duration::from_millis(5));
+        let task = AsyncTask::new_with_timeout(Duration::from_millis(5), pending::<()>());
         let (fut, mut rx) = task.into_parts();
 
         assert_eq!(None, rx.try_recv());
@@ -332,10 +366,15 @@ mod test {
             Ok(JsValue::NULL)
         }))
         .await
-        .unwrap();
+        .unwrap_or_else(|e| {
+            panic!("awaiting promise failed: {e:?}");
+        });
 
         // Spawn
-        let v = rx.try_recv().expect("future loaded no value");
+        let v = rx.try_recv().unwrap_or_else(|| {
+            panic!("expected result after await");
+        });
+        assert!(matches!(v, Err(TaskError::Timeout(_))), "");
         assert!(v.is_err(), "timeout should have triggered!");
     }
 }